requirements.txt
feedparser
youtube-transcript-api
google-generativeai
python-dotenv
